SimulationInfo.* files renamed to StandardTestInfo.*
